task: refactor levels page
